onboarding-agent/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
google-genai
google-adk
pydantic
python-multipart

